refactoring, merge all super admin php func files to func.php
